requirements.txt
streamlit  
sentence_transformers
crewai
einops
PyPDF2
chromadb
python-dotenv
dashscope
openai
